Feat/handoff 03 mvp backend

+++ app/loans/schemas.py
@@ -1,5 +1,4 @@
 from datetime import datetime
-from decimal import Decimal
 from uuid import UUID
 
 from pydantic import BaseModel, Field, model_validator
@@ -7,7 +6,7 @@ from pydantic import BaseModel, Field, model_validator
 
 class LoanApplicationCreate(BaseModel):
     business_id: UUID
-    requested_amount: Decimal
+    requested_amount: int
     purpose: str | None = None
     repayment_preference: str | None = None
 
@@ -15,7 +14,7 @@ class LoanApplicationCreate(BaseModel):
 class LoanApplicationOut(BaseModel):
     id: UUID
     project_id: UUID
-    requested_amount: Decimal
+    requested_amount: int
     purpose: str | None
     repayment_preference: str | None
     status: str
