+++ app/market/schemas.py
@@ -1,4 +1,3 @@
-from decimal import Decimal
 from uuid import UUID
 
 from pydantic import BaseModel, ConfigDict, Field
@@ -9,17 +8,17 @@ class ListingCreate(BaseModel):
     # Same missing-bound class as OrderCreate.amount below: the service checks
     # `min_ticket <= 0` but never bounded target_amount, so a zero or negative
     # target produced a listing no order could ever fill (remaining <= 0).
-    target_amount: Decimal = Field(gt=0)
-    min_ticket: Decimal = Field(gt=0)
+    target_amount: int = Field(gt=0)
+    min_ticket: int = Field(gt=0)
 
 
 class ListingOut(BaseModel):
     id: UUID
     contract_id: UUID
-    target_amount: Decimal
-    min_ticket: Decimal | None
+    target_amount: int
+    min_ticket: int | None
     status: str
-    funded_amount: Decimal
+    funded_amount: int
 
     model_config = {"from_attributes": True}
 
@@ -32,14 +31,14 @@ class OrderCreate(BaseModel):
     # (the column is nullable), and `amount > remaining` cannot catch a negative
     # — so without this a negative order would be accepted, decrementing the
     # listing's funded_amount and writing a FILLED order for negative money.
-    amount: Decimal = Field(gt=0)
+    amount: int = Field(gt=0)
     ack_risk_disclosure: bool = Field(default=False, alias="ackRiskDisclosure")
 
 
 class OrderOut(BaseModel):
     id: UUID
     listing_id: UUID
-    amount: Decimal
+    amount: int
     status: str
 
     model_config = {"from_attributes": True}
